refactored according to PR suggestions
