+++ lib/helpers/find-element-strategy.js
@@ -4,12 +4,13 @@ module.exports = function findStrategy() {
 
     var value = arguments[0],
         relative = (arguments.length > 1 ? arguments[1] : false),
-        using, query, tag, classOrId, classOrIdName;
+        using, query, tag, classOrId, classOrIdName, xpathPrefix;
 
-    if (relative)
+    if (relative) {
         xpathPrefix = './';
-    else
+    } else {
         xpathPrefix = '//';
+    }
 
     if (typeof value !== 'string') {
         throw new ErrorHandler.ProtocolError('selector needs to be typeof `string`');
