+++ lib/protocol/elementIdElements.js
@@ -24,7 +24,7 @@ module.exports = function elementIdElements (id, selector) {
 
     var findStrategy = require('../helpers/find-element-strategy.js');
 
-    var found = findStrategy(selector);
+    var found = findStrategy(selector, true);
 
     var requestPath = '/session/:sessionId/element/:id/elements';
     requestPath = requestPath.replace(/:id/gi, id);
