+++ test/site/www/index.html
@@ -88,6 +88,11 @@
         <div class="visibletest hidden">I am not visible</div>
         <div class="invisibleParent hidden"><div class="visibletestInner">I am not visible, am I?</div></div>
         <hr>
+        <table>
+            <tr><td>2015-03-01</td><td>12:00</td></tr>
+            <tr><td>2015-03-02</td><td>12:00</td></tr>
+        </table>
+        <hr>
         <form method="GET" action="index.html" class="send">
             <input name="a" value="a">
             <input name="b" value="b">
